+++ TigerCave/src/main/java/com/sparta/tigercave/security/UserDetailImpl.java
@@ -6,6 +6,7 @@ import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.Collection;
 
@@ -13,12 +14,12 @@ public class UserDetailImpl implements UserDetails {
 
     private final User user;
 
-    private final Long userId;
+    private final BigInteger userId;
     private final String username;
 
     private final String password;
 
-    public UserDetailImpl(User user, Long userId, String username, String password){
+    public UserDetailImpl(User user, BigInteger userId, String username, String password){
         this.user = user;
         this.userId = userId;
         this.username = username;
@@ -43,7 +44,7 @@ public class UserDetailImpl implements UserDetails {
         return authorities;
     }
 
-    public Long getUserId() {
+    public BigInteger getUserId() {
         return this.userId;
     }
